### 2016-12-08 #134 @whedon
[PRE REVIEW]: Optlang: A Python interface to common mathematical optimization solvers
**Submitting author:** @KristianJensen (<a href="http://orcid.org/0000-0002-2796-805X">Kristian Jensen</a>)
**Repository:** <a href="https://github.com/biosustain/optlang" target ="_blank">https://github.com/biosustain/optlang</a>
**Version:** v1.0.0

Thanks for submitting your paper to JOSS KristianJensen. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
