### 2016-12-08 #134 @arfon
@kyleniemeyer @labarba would either of you be wiling to edit this submission for JOSS?
