### 2016-12-08 #134 @kyleniemeyer
I can handle this one, unless @labarba prefers to.
