### 2016-12-14 #134 @kyleniemeyer
Hi @jordigh @jkahn would either of you be interested in helping review this?

I don't see any current reviewers with explicit expertise in optimization, so I sent out a request on Twitter: https://twitter.com/kyleniemeyer/status/808837158714806272

@KristianJensen any suggestions you have on possible domain-expert reviewers would be appreciated as well; I try to get two people, including one area expert, to take a look.
