### 2016-12-14 #134 @moorepants
@kyleniemeyer I'd be interested in reviewing this. Can anyone review?
