### 2016-12-14 #134 @kyleniemeyer
@moorepants absolutely! I was hoping to get someone familiar with SymPy, since this package relies on it. Please take a look at the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines), and add yourself as a reviewer via a PR to the [reviewer csv file](https://github.com/openjournals/joss/blob/master/docs/reviewers.csv).
