### 2016-12-14 #139 @whedon
[REVIEW]: Optlang: A Python interface to common mathematical optimization solvers
**Submitting author:** @KristianJensen (<a href="http://orcid.org/0000-0002-2796-805X">Kristian Jensen</a>)
**Repository:** <a href="https://github.com/biosustain/optlang" target ="_blank">https://github.com/biosustain/optlang</a>
**Version:** v1.0.0

- [x] **Authorship:** Has the submitting author (KristianJensen) made major contributions to the software?
