### 2016-12-14 biosustain/optlang#54 @moorepants
[JOSS Review] Installation issues
On the first try installing I get a print statement that says "No installers available." on module import:

```
moorepants@garuda:~$ conda create -n optlang-review ipython
Fetching package metadata ...........
Solving package specifications: ..........

Package plan for installation in environment /home/moorepants/miniconda3/envs/optlang-review:

The following NEW packages will be INSTALLED:

    decorator:        4.0.10-py35_1
    ipython:          5.1.0-py35_0
    ipython_genutils: 0.1.0-py35_0
    openssl:          1.0.2j-0
    path.py:          8.2.1-py35_0
    pexpect:          4.0.1-py35_0
    pickleshare:      0.7.4-py35_0
    pip:              9.0.1-py35_1
    prompt_toolkit:   1.0.9-py35_0
    ptyprocess:       0.5.1-py35_0
    pygments:         2.1.3-py35_0
    python:           3.5.2-0
    readline:         6.2-2
    setuptools:       27.2.0-py35_0
    simplegeneric:    0.8.1-py35_1
    six:              1.10.0-py35_0
    sqlite:           3.13.0-0
    tk:               8.5.18-0
    traitlets:        4.3.1-py35_0
    wcwidth:          0.1.7-py35_0
    wheel:            0.29.0-py35_0
    xz:               5.2.2-0
    zlib:             1.2.8-3

Linking packages ...
[      COMPLETE      ]|############################################################################| 100%
#
# To activate this environment, use:
# > source activate optlang-review
#
# To deactivate this environment, use:
# > source deactivate optlang-review
#

moorepants@garuda:~$ act optlang-review
(optlang-review) moorepants@garuda:~$ pip install optlang
Collecting optlang
  Using cached optlang-1.0.0-py2.py3-none-any.whl
Requirement already satisfied: six>=1.9.0 in ./miniconda3/envs/optlang-review/lib/python3.5/site-packages (from optlang)
Collecting sympy>=1.0.0 (from optlang)
Collecting mpmath>=0.19 (from sympy>=1.0.0->optlang)
Installing collected packages: mpmath, sympy, optlang
Successfully installed mpmath-0.19 optlang-1.0.0 sympy-1.0
(optlang-review) moorepants@garuda:~$ ipython -c "import optlang"
No solvers available.
(optlang-review) moorepants@garuda:~$ ipython
Python 3.5.2 |Continuum Analytics, Inc.| (default, Jul  2 2016, 17:53:06)
Type "copyright", "credits" or "license" for more information.

IPython 5.1.0 -- An enhanced Interactive Python.
?         -> Introduction and overview of IPython's features.
%quickref -> Quick reference.
help      -> Python's own help system.
object?   -> Details about 'object', use 'object??' for extra details.

In [1]: import optlang
No solvers available.

In [2]:
Do you really want to exit ([y]/n)?
(optlang-review) moorepants@garuda:~$ python -c "import optlang"
No solvers available.
```

I'd suggest having all required dependencies installed automatically via the pip dependency resolution. Also, shouldn't `import optlang` raise and ImportError if the dependencies are not installed?
