### 2016-12-14 biosustain/optlang#54 @moorepants
After pip installing swiglpk the example runs:

```
(optlang-review) moorepants@garuda:~$ pip install swiglpk
Collecting swiglpk
  Downloading swiglpk-1.3.3-cp35-cp35m-manylinux1_x86_64.whl (1.8MB)
    100% |████████████████████████████████| 1.8MB 549kB/s
Installing collected packages: swiglpk
Successfully installed swiglpk-1.3.3
(optlang-review) moorepants@garuda:~$ python -c 'import optlang'
(optlang-review) moorepants@garuda:~$ ipython
Python 3.5.2 |Continuum Analytics, Inc.| (default, Jul  2 2016, 17:53:06)
Type "copyright", "credits" or "license" for more information.

IPython 5.1.0 -- An enhanced Interactive Python.
?         -> Introduction and overview of IPython's features.
%quickref -> Quick reference.
help      -> Python's own help system.
object?   -> Details about 'object', use 'object??' for extra details.

In [1]: %paste
from optlang.glpk_interface import Model, Variable, Constraint, Objective

x1 = Variable('x1', lb=0)
x2 = Variable('x2', lb=0)
x3 = Variable('x3', lb=0)

c1 = Constraint(x1 + x2 + x3, ub=100)
c2 = Constraint(10 * x1 + 4 * x2 + 5 * x3, ub=600)
c3 = Constraint(2 * x1 + 2 * x2 + 6 * x3, ub=300)

obj = Objective(10 * x1 + 6 * x2 + 4 * x3, direction='max')

model = Model(name='Simple model')
model.objective = obj
model.add([c1, c2, c3])

status = model.optimize()

## -- End pasted text --

In [2]: print('status:', model.status)
status: optimal

In [3]: print('objective value:', model.objective.value)
objective value: 733.3333333333333

In [4]: for var_name, var in model.variables.iteritems():
   ...:     print(var_name, '=', var.primal)
   ...:
x1 = 33.33333333333333
x2 = 66.66666666666667
x3 = 0.0
```

The example uses old style Python printing in the README. I'd suggest updating to Python 2.7/3+ valid code.
