### 2016-12-14 biosustain/optlang#54 @moorepants
I see that you have 4 optional dependcies for solvers. I removed swiggplk and tried the scipy backend, but get this error:

```
(optlang-review) moorepants@garuda:~$ pip uninstall swiglpk
Uninstalling swiglpk-1.3.3:
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk-1.3.3.dist-info/DESCRIPTION.rst
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk-1.3.3.dist-info/INSTALLER
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk-1.3.3.dist-info/METADATA
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk-1.3.3.dist-info/RECORD
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk-1.3.3.dist-info/WHEEL
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk-1.3.3.dist-info/metadata.json
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk/.libs/libglpk-190a7631.so.40.1.0
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk/__init__.py
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk/__pycache__/__init__.cpython-35.pyc
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk/__pycache__/swiglpk.cpython-35.pyc
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk/_swiglpk.cpython-35m-x86_64-linux-gnu.so
  /home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/swiglpk/swiglpk.py
Proceed (y/n)? y
  Successfully uninstalled swiglpk-1.3.3
(optlang-review) moorepants@garuda:~$ conda install scipy
Fetching package metadata ...........
Solving package specifications: ..........

Package plan for installation in environment /home/moorepants/miniconda3/envs/optlang-review:

The following NEW packages will be INSTALLED:

    libgfortran: 3.0.0-1
    mkl:         11.3.3-0
    numpy:       1.11.2-py35_0
    scipy:       0.18.1-np111py35_0

Linking packages ...
[      COMPLETE      ]|############################################################################| 100%
(optlang-review) moorepants@garuda:~$ ipython
Python 3.5.2 |Continuum Analytics, Inc.| (default, Jul  2 2016, 17:53:06)
Type "copyright", "credits" or "license" for more information.

IPython 5.1.0 -- An enhanced Interactive Python.
?         -> Introduction and overview of IPython's features.
%quickref -> Quick reference.
help      -> Python's own help system.
object?   -> Details about 'object', use 'object??' for extra details.

In [1]: %paste
from optlang.glpk_interface import Model, Variable, Constraint, Objective

x1 = Variable('x1', lb=0)
x2 = Variable('x2', lb=0)
x3 = Variable('x3', lb=0)

c1 = Constraint(x1 + x2 + x3, ub=100)
c2 = Constraint(10 * x1 + 4 * x2 + 5 * x3, ub=600)
c3 = Constraint(2 * x1 + 2 * x2 + 6 * x3, ub=300)

obj = Objective(10 * x1 + 6 * x2 + 4 * x3, direction='max')

model = Model(name='Simple model')
model.objective = obj
model.add([c1, c2, c3])

status = model.optimize()

## -- End pasted text --
ImportError                               Traceback (most recent call last)
<ipython-input-1-32194ddcbb11> in <module>()
----> 1 from optlang.glpk_interface import Model, Variable, Constraint, Objective
      2
      3 x1 = Variable('x1', lb=0)
      4 x2 = Variable('x2', lb=0)
      5 x3 = Variable('x3', lb=0)

/home/moorepants/miniconda3/envs/optlang-review/lib/python3.5/site-packages/optlang/glpk_interface.py in <module>()
     41 log = logging.getLogger(__name__)
     42
---> 43 from swiglpk import glp_find_col, glp_get_col_prim, glp_get_col_dual, GLP_CV, GLP_IV, GLP_BV, GLP_UNDEF, GLP_FEAS, \
     44     GLP_INFEAS, GLP_NOFEAS, GLP_OPT, GLP_UNBND, \
     45     glp_set_col_kind, glp_find_row, glp_get_row_prim, glp_get_row_dual, glp_get_obj_val, glp_set_obj_dir, glp_init_smcp, \

ImportError: No module named 'swiglpk'
```

Should any of the backends work?
