### 2016-12-14 biosustain/optlang#54 @moorepants
The installation instructions on readthedocs seem to be more up-to-date and clearer than the README. It would be a good idea to have them match in terms of content.
