### 2016-12-14 #139 @moorepants
Some initial notes:

- The installation was a bit fuzzy on the dependencies. I think that at least scipy and swiglpk should be pulled in automatically with install, see https://github.com/biosustain/optlang/issues/54 for more.
- It seems the JOSS submission author, @KristianJensen, is more of a [minor contributor of the software](https://github.com/biosustain/optlang/graphs/contributors). It seems the largest contribution is by far by @phantomas1234. Is there are reason they are not the primary author?
