### 2016-12-14 95ae9da
chore: make example in README solver agnostic
And add a note on how to use a specific solver
