### 2016-12-14 21189c9
docs: add comments to README code
