### 2016-12-14 e8688b8
docs: update README to reflect swiglpk as a dep.
swiglpk is a hard dependency now, so that 'pip install optlang' works out of the box.
