### 2016-12-14 ab8881d
docs: update install instructions
Swiglpk is now a hard dependency so the installation instructions are updated accordingly.
