### 2016-12-14 #139 @KristianJensen
@moorepants Thanks for your comments so far. Regarding the contributions:
Due to a mistake on my part (forgetting to set my email in git) that was recently corrected, the majority of my commits are not linked to my github user. Doing `git log | grep 'Author: Kristian' | wc` in the current master branch returns 115 commits authored by me.

The project was originally initiated by @phantomas1234 who also drafted several of the solver interfaces and core functionality. I have contributed by adding several major features (e.g. QP support, scipy_interface, JSON-storage) and by shaping up the code from a "working draft" to "ready for submission". This has included many refactorings and bug fixes, unifying API and behaviour across solvers and setting up a common unittest suite for all the supported solvers.
Based on this we decided that I would be the submitting author, and the other authors agree that my contributions are well within the definition of "major".
