### 2016-12-14 biosustain/optlang#54 @KristianJensen
Swiglpk is now a requirement, so after installing optlang with pip, the package can be used for optimizing problems without any further installations.
Since scipy is a huge package, and it's linear programming functionalities are quite limited, we do not think it makes sense to also have this as a mandatory dependency.

The reason that "No solvers available" does not raise an error is that optlang can still be used to formulate, manipulate and save problems even without a solver. With swiglpk as a requirement, this is not so relevant however.

Your comments about the README were very valid and have been addressed (in #55). The example in the README should work with other solvers now.
