### 2016-12-14 #139 @phantomas1234
I copy all of that. Thank you for your review and your valuable comments @moorepants. We could fix the commit history for github to report @KristianJensen's contributions accurately but that would be fairly destructive and we would like to avoid that.
