### 2016-12-14 #139 @moorepants
Thanks for the clarification on the contributions. That's acceptable to me. No need to fix the git history. I assume that JOSS somehow includes all contributors as authors. So it should be fine.
