### 2016-12-14 #139 @kyleniemeyer
@moorepants @phantomas1234 @KristianJensen I'll jump in here and state that JOSS does not require the lead/submitting author to be the person who has made the majority of commits/contributions to the original code.
