### 2016-12-14 #139 @moorepants
My comments on contribution arise based on this checkbox: "Authorship: Has the submitting author (KristianJensen) made major contributions to the software?". When i looked at the Github contributors graph and commit list, it wasn't clear to me that this was satisfied. Can JOSS be more clear on what "major contributions" means?

Also @kyleniemeyer, can you clarify who JOSS assigns as authors to the "paper"?
