### 2016-12-14 7550ea5
docs: update dependency versions in readme
