+++ README.rst
@@ -92,9 +92,9 @@ Development
 
 The following dependencies are needed.
 
--  `sympy >= 0.7.5 <http://sympy.org/en/index.html>`__
+-  `sympy >= 1.0.0 <http://sympy.org/en/index.html>`__
 -  `six >= 1.9.0 <https://pypi.python.org/pypi/six>`__
--  `swiglpk >= 0.1.0 <https://pypi.python.org/pypi/swiglpk>`__
+-  `swiglpk >= 1.3.0 <https://pypi.python.org/pypi/swiglpk>`__
 
 The following are optional dependencies that allow other solvers to be used.
 
