### 2016-12-14 #139 @moorepants
Additional items to address:

- [ ] State who the target audience is in the documentation statement of need.
- [ ] I recommend adding a CONTRIBUTING file that mirrors and maybe enhances this statement: http://optlang.readthedocs.io/en/latest/developers.html#contribute. You could be a little more explicit on where to seek support.
- [ ] Also in paper.md "who the target audience is".
