### 2016-12-14 #139 @moorepants
I finished going through the initial look. I've tested the software (automated and examples). It's super cool. I think this is an excellent use of SymPy and very useful for solving many problems. Great work!

If you can address those concerns above, I'm happy to approve this.
