### 2016-12-14 #139 @kyleniemeyer
@moorepants fair point. We don't really have an explicit definition of "major contribution"; it's mainly there to ensure we don't have someone submitting that isn't a contributor at all (and in any event the major contributors should certainly make up the author list). In this particular case, your question was certainly fine, but it seems the situation is fine as far as we're concerned.

We leave it up to the project contributors to decide who are the paper authors, but from my perspective the contributor list and author list should mostly match. @arfon do you have anything to add here?
