### 2016-12-15 #139 @arfon
> We leave it up to the project contributors to decide who are the paper authors, but from my perspective the contributor list and author list should mostly match. @arfon do you have anything to add here?

That's correct, we don't have any hard and fast rules here. I too would expect these to mostly match but I'm sure there are scenarios where they won't.
