### 2016-12-15 570fcf2
docs: update contributing guidelines
And add CONTRIBUTIONS.md file
