### 2016-12-15 #139 @andim
Regarding the point "statement of need" I have two suggestions:
- Point out more prominently in the documentation/README that optlang is a modeling language for optimization problems (as opposed to a package providing solvers)
- Feature the range of optimization problems supported (seems to be LP/QP/MILP currently) more prominently e.g. this info is missing in the README
- Discuss benefits/drawbacks relative to the more standard `scipy.optimize`, which not only provides a range of solvers but also a generic interface to them (`scipy.optimize.minimize`)
