### 2016-12-19 317d9c4
docs: update readme, paper and documentation
Add target audience to paper and docs/index
Add solver problem capabilities to readme
Move installation issue from readme to docs/installation
