### 2016-12-19 b06eff8
test: remove non-existing libglp package
