### 2016-12-19 a291e06
feat: implement set_linear_coefficients in scipy_interface
This was missing in scipy
