### 2016-12-19 #139 @KristianJensen
@moorepants We have now clarified who the target audience is, both in the paper and documentation and added a CONTRIBUTING.md file.

Furthermore as per the suggestions above (thanks @andim !) we have made it more clear that optlang is a modeling language (including a new paper title) and added currently supported solver capabilities to the readme. However we would prefer not to have lengthy comparisons of optlang and other available software in the documentation, although with the clarified description of optlang as a modeling language, it should be more clear what distinguishes it from e.g. `scipy.optimize.minimize`.
