### 2016-12-19 711e1a2
fix: all the bugs introduced by adding set_linear_coefficients
