### 2016-12-19 #139 @moorepants
Thanks. All looks good to me now. I recommend for acceptance.
