### 2016-12-20 b1ae253
test: add tests for set_linear_coefficients
