### 2016-12-20 24a8768
fix: use float division
