+++ optlang/tests/abstract_test_cases.py
@@ -13,6 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from __future__ import division
+
 import abc
 import unittest
 
