### 2016-12-20 78175e8
fix: handle if _problem is not set
