+++ optlang/interface.py
@@ -443,7 +443,7 @@ class OptimizationExpression(object):
     @property
     def problem(self):
         """A reference to the model that the object belongs to (or None)"""
-        return self._problem
+        return getattr(self, '_problem', None)
 
     @problem.setter
     def problem(self, value):
