+++ optlang/scipy_interface.py
@@ -433,9 +433,9 @@ class Constraint(interface.Constraint):
 
 @six.add_metaclass(inheritdocstring)
 class Objective(interface.Objective):
-    def __init__(self, *args, **kwargs):
-        super(Objective, self).__init__(*args, **kwargs)
-        if not self.is_Linear:
+    def __init__(self, expression, sloppy=False, **kwargs):
+        super(Objective, self).__init__(expression, sloppy, **kwargs)
+        if not (sloppy or self.is_Linear):
             raise ValueError(
                 "Scipy only supports linear objectives. %s is not linear." % self)
 
