### 2016-12-20 9787bec
fix: correct expression parsing
1.0 * x**2 should not be parsed to (x**2: 1.0)
