### 2016-12-20 d607d21
Fix bugs (#56)
* fix: change cplex solutiontarget to optimalitytarget

Cplex 12.7 has replaced solutiontarget with optimalitytarget. This fix
tries to use optimalitytarget and falls back on solutiontarget if
optimalitytarget doesn’t exist.

* feat: implement set_linear_coefficients in scipy_interface

This was missing in scipy

* fix: all the bugs introduced by adding set_linear_coefficients

* fix: typo

* fix: flake8

* test: add tests for set_linear_coefficients

* fix: flake8

* fix: use float division

* fix: handle if _problem is not set

* chore: add sloppy to scipy_interface objective

[…]
