+++ optlang/scipy_interface.py
@@ -33,6 +33,7 @@ import six
 from optlang import interface
 from scipy.optimize import linprog
 from optlang.util import inheritdocstring
+import sympy
 
 SCIPY_STATUS = {
     0: interface.OPTIMAL,
@@ -403,27 +404,49 @@ class Constraint(interface.Constraint):
                 self.problem.problem.set_constraint_bound(self.upper_constraint_name, value)
         self._ub = value
 
-    def coefficient_dict(self):
+    def coefficient_dict(self, names=True):
         if self.expression.is_Add:
-            coefficient_dict = {variable.name: coef for variable, coef in
+            coefficient_dict = {variable: coef for variable, coef in
                                 self.expression.as_coefficients_dict().items()}
         elif self.expression.is_Atom and self.expression.is_Symbol:
-            coefficient_dict = {self.expression.name: 1}
+            coefficient_dict = {self.expression: 1}
         elif self.expression.is_Mul and len(self.expression.args) <= 2:
             args = self.expression.args
             coefficient_dict = {args[1]: float(args[0])}
+        elif self.expression.is_Number:
+            coefficient_dict = {}
         else:
-            raise ValueError("Invalid expression")
+            raise ValueError("Invalid expression: " + str(self.expression))
+        if names is True:
+            coefficient_dict = {var.name: coef for var, coef in coefficient_dict.items()}
         return coefficient_dict
+
+    def set_linear_coefficients(self, coefficients):
+        lb, ub = self.lb, self.ub
+        self.lb, self.ub = None, None
+        coefficients_dict = self.coefficient_dict(names=False)
+        coefficients_dict.update(coefficients)
+        self._expression = sympy.Add(*(v * k for k, v in coefficients_dict.items()))
+        self.lb = lb
+        self.ub = ub
 
 
 @six.add_metaclass(inheritdocstring)
 class Objective(interface.Objective):
-    def __init__(self, *args, **kwargs):
-        super(Objective, self).__init__(*args, **kwargs)
-        if not self.is_Linear:
+    def __init__(self, expression, sloppy=False, **kwargs):
+        super(Objective, self).__init__(expression, sloppy, **kwargs)
+        if not (sloppy or self.is_Linear):
             raise ValueError(
                 "Scipy only supports linear objectives. %s is not linear." % self)
+
+    def _get_expression(self):
+        if self.problem is not None:
+            coefficients_dict = self.problem.problem.objective
+            coefficients_dict = {
+                self.problem._variables[name]: coef for name, coef in coefficients_dict.items() if name in self.problem._variables
+            }
+            self._expression = sympy.Add(*(v * k for k, v in coefficients_dict.items()))
+        return self._expression
 
     @property
     def value(self):
@@ -440,19 +463,22 @@ class Objective(interface.Objective):
 
     def coefficient_dict(self):
         if self.expression.is_Add:
-            coefficient_dict = {variable.name: coef for variable, coef in
+            coefficient_dict = {variable: coef for variable, coef in
                                 self.expression.as_coefficients_dict().items()}
-        elif self.expression.is_Atom:
-            if self.expression.is_Symbol:
-                coefficient_dict = {self.expression.name: 1}
-            else:
-                coefficient_dict = {}
-        elif self.expression.is_Mul and len(self.expression.args) <= 2:
+        elif self.expression.is_Atom and self.expression.is_Symbol:
+            coefficient_dict = {self.expression: 1}
+        elif self.expression.is_Mul and len(self.expression.args) <= 2 and self.expression.args[1].is_Symbol:
             args = self.expression.args
-            coefficient_dict = {args[1].name: float(args[0])}
+            coefficient_dict = {args[1]: float(args[0])}
+        elif self.expression.is_Number:
[…]
