### 2016-12-23 152c2d1
Some minor writing fixes
I made some minor writing fixes, to be adopted before acceptance in JOSS (https://github.com/openjournals/joss-reviews/issues/139)
