### 2016-12-23 #139 @kyleniemeyer
Thanks for the review @moorepants!

@KristianJensen I just made a few minor edits to the paper in https://github.com/biosustain/optlang/pull/57, please take a look and then merge those. After that, I think we will be good to go for acceptance in JOSS—you can then archive the software, and please provide the associated DOI here.
