### 2017-01-02 be7d226
Some minor writing fixes (#57)
I made some minor writing fixes, to be adopted before acceptance in JOSS (https://github.com/openjournals/joss-reviews/issues/139)
