### 2017-01-02 52a632e
Joss review (#55)
* chore: make example in README solver agnostic

And add a note on how to use a specific solver

* docs: add comments to README code

* Update requirements.txt

* docs: update README to reflect swiglpk as a dep.

swiglpk is a hard dependency now, so that 'pip install optlang' works out of the box.

* docs: update install instructions

Swiglpk is now a hard dependency so the installation instructions are updated accordingly.

* docs: update dependency versions in readme

* docs: update contributing guidelines

And add CONTRIBUTIONS.md file

* docs: update readme, paper and doc
[…]
