+++ CONTRIBUTING.md
@@ -1,0 +1,7 @@
+Contribute
+==========
+
+Contributions to optlang are very welcome. Fork optlang at `github
+<http://github.com/biosustain/optlang>`_, implement your feature and send us
+a pull request. Also, please use the GitHub `issue tracker <https://github.com/biosustain/optlang/issues>`_
+to let us know about bugs or feature requests, or if you have problems or questions regarding optlang.
