+++ README.rst
@@ -10,12 +10,15 @@ problems, i.e. maximizing or minimizing an objective function over a set
 of variables subject to a number of constraints. Optlang provides a
 common interface to a series of optimization tools, so different solver
 backends can be changed in a transparent way.
-Optlang takes advantage of the symbolic math library
+Optlang's object-oriented API takes advantage of the symbolic math library
 `sympy <http://sympy.org/en/index.html>`__ to allow objective functions
 and constraints to be easily formulated from symbolic expressions of
 variables (see examples).
 
 Show us some love by staring this repo if you find optlang useful!
+
+Also, please use the GitHub `issue tracker <https://github.com/biosustain/optlang/issues>`_
+to let us know about bugs or feature requests, or if you have problems or questions regarding optlang.
 
 Installation
 ~~~~~~~~~~~~
@@ -26,13 +29,25 @@ Install using pip
 
     pip install optlang
 
-Then you could install `swiglpk <https://github.com/biosustain/swiglpk>`_
+This will also install `swiglpk <https://github.com/biosustain/swiglpk>`_, an interface to the open source (mixed integer) LP solver `GLPK <https://www.gnu.org/software/glpk/>`_.
+Quadratic programming (and MIQP) is supported through additional optional solvers (see below).
 
-::
+Dependencies
+~~~~~~~~~~~~
 
-    pip install swiglpk
+The following dependencies are needed.
 
-to solve your optimization problems using `GLPK <https://www.gnu.org/software/glpk/>`_ (see below for further supported solvers).
+-  `sympy >= 1.0.0 <http://sympy.org/en/index.html>`__
+-  `six >= 1.9.0 <https://pypi.python.org/pypi/six>`__
+-  `swiglpk >= 1.3.0 <https://pypi.python.org/pypi/swiglpk>`__
+
+The following are optional dependencies that allow other solvers to be used.
+
+-  `cplex <https://www-01.ibm.com/software/commerce/optimization/cplex-optimizer/>`__ (LP, MILP, QP, MIQP)
+-  `gurobipy <http://www.gurobi.com>`__ (LP, MILP (QP and MIQP support will be added in the future))
+-  `scipy <http://www.scipy.org>`__ (LP)
+
+
 
 Example
 ~~~~~~~
@@ -40,21 +55,25 @@ Example
 Formulating and solving the problem is straightforward (example taken
 from `GLPK documentation <http://www.gnu.org/software/glpk>`__):
 
-::
+.. code-block:: python
 
     from __future__ import print_function
-    from optlang.glpk_interface import Model, Variable, Constraint, Objective
+    from optlang import Model, Variable, Constraint, Objective
 
+    # All the (symbolic) variables are declared, with a name and optionally a lower and/or upper bound.
     x1 = Variable('x1', lb=0)
     x2 = Variable('x2', lb=0)
     x3 = Variable('x3', lb=0)
 
+    # A constraint is constructed from an expression of variables and a lower and/or upper bound (lb and ub).
     c1 = Constraint(x1 + x2 + x3, ub=100)
     c2 = Constraint(10 * x1 + 4 * x2 + 5 * x3, ub=600)
     c3 = Constraint(2 * x1 + 2 * x2 + 6 * x3, ub=300)
 
+    # An objective can be formulated
     obj = Objective(10 * x1 + 6 * x2 + 4 * x3, direction='max')
 
+    # Variables, constraints and objective are combined in a Model object, which can subsequently be optimized.
     model = Model(name='Simple model')
     model.objective = obj
     model.add([c1, c2, c3])
@@ -63,6 +82,7 @@ from `GLPK documentation <http://www.gnu.org/software/glpk>`__):
 
[…]
