+++ requirements.txt
@@ -1,2 +1,3 @@
 sympy>=1.0.0
 six>=1.9.0
+swiglpk>=1.3.0
