### 2017-01-03 #139 @KristianJensen
Hi @kyleniemeyer

Everything is now merged, released and archived on zenodo at https://doi.org/10.5281/zenodo.228165
