### 2017-01-03 #139 @arfon
@kyleniemeyer - is this good to accept now?
